+++ noodles-vcf/src/header/format.rs
@@ -1,5 +1,3 @@
 //! VCF header genotype format record components.
 
 pub mod key;
-
-pub use self::key::Key;