+++ noodles-vcf/src/record/genotypes/sample/value.rs
@@ -159,8 +159,8 @@ impl Value {
     ///
     /// ```
     /// use noodles_vcf::{
-    ///     header::{format::key, record::value::{map::Format, Map}},
-    ///     record::genotypes::sample::Value,
+    ///     header::{record::value::{map::Format, Map}},
+    ///     record::genotypes::{keys::key, sample::Value},
     /// };
     ///
     /// let format = Map::<Format>::from(&key::CONDITIONAL_GENOTYPE_QUALITY);