+++ noodles-vcf/src/record/parser.rs
@@ -44,12 +44,9 @@ mod tests {
 
     #[test]
     fn test_parse_with_genotype_info() -> Result<(), Box<dyn std::error::Error>> {
-        use crate::{
-            header::format::key,
-            record::{
-                genotypes::{sample::Value, Keys},
-                Genotypes,
-            },
+        use crate::record::{
+            genotypes::{keys::key, sample::Value, Keys},
+            Genotypes,
         };
 
         let header = Header::builder().add_sample_name("sample0").build();